lvtn_project/requirements.txt
Django
djangorestframework
beautifulsoup4
psycopg2
django-filter
Pillow
django-rest-swagger
djangorestframework_simplejwt
django-summernote
django-cors-headers